# src/RedShirt.Adventure.Realm/Models/Characters/CharacterResources/CharacterResourcesPutRequest.cs
namespace RedShirt.Adventure.Realm.Models.Characters.CharacterResources;

public class CharacterResourcesPutRequest
{
    public ulong HealthCurrent { get; init; }
    public ulong HealthMaximum { get; init; }
}