# swagger/RedShirt.Adventure.Realm.Interop/swagger.json
{
...
  "openapi": "3.0.0",
  "info": {
    "title": "RedShirt.Adventure.Realm",
    "version": "1.0.0"
  },
  "paths": {
...
    "/character": {
      "get": {
        "tags": [
          "Character"
        ],
        "operationId": "Character_GetList",
        "parameters": [
          {
            "name": "accountId",
            "in": "query",
            "schema": {
              "type": "string",
              "nullable": true
            },
            "x-position": 1
          },
          {
            "name": "firstName",
            "in": "query",
            "schema": {
              "type": "string",
              "nullable": true
            },
            "x-position": 2
          },
          {
            "name": "lastName",
            "in": "query",
            "schema": {
              "type": "string",
              "nullable": true
            },
            "x-position": 3
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 4
          },
          {
            "name": "continuationToken",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "guid",
              "nullable": true
            },
            "x-position": 5
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CharacterSearchResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Character"
        ],
        "operationId": "Character_Post",
        "requestBody": {
          "x-name": "postRequest",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CharacterPostRequest"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CharacterDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/character/{id}/resources": {
      "delete": {
        "tags": [
          "CharacterResources"
        ],
        "operationId": "CharacterResources_Delete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "CharacterResources"
        ],
        "operationId": "CharacterResources_Get",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CharacterResourcesDto"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "CharacterResources"
        ],
        "operationId": "CharacterResources_Patch",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          }
        ],
        "requestBody": {
          "x-name": "patchRequest",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CharacterResourcesPatchRequest"
              }
            }
          },
          "required": true,
          "x-position": 2
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CharacterResourcesDto"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "304": {
            "description": ""
          }
        }
      },
      "put": {
        "tags": [
          "CharacterResources"
        ],
        "operationId": "CharacterResources_Put",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          }
        ],
        "requestBody": {
          "x-name": "putRequest",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CharacterResourcesPutRequest"
              }
            }
          },
          "required": true,
          "x-position": 2
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CharacterResourcesDto"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "304": {
            "description": ""
          }
        }
      }
    },
    "/character/attributes/resources": {
      "get": {
        "tags": [
          "CharacterResources"
        ],
        "operationId": "CharacterResources_List",
        "parameters": [
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 1
          },
          {
            "name": "continuationToken",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "guid",
              "nullable": true
            },
            "x-position": 2
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CharacterResourcesSearchResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ProblemDetails": {
        "type": "object",
        "additionalProperties": {
          "nullable": true
        },
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "CharacterDto": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "format": "guid"
          },
          "createdAtUtc": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAtUtc": {
            "type": "string",
            "format": "date-time"
          },
          "accountId": {
            "type": "string"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          }
        }
      },
      "CharacterSearchResponse": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "continuationToken": {
            "type": "string",
            "format": "guid",
            "nullable": true
          },
          "records": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CharacterDto"
            }
          }
        }
      },
...
      "CharacterPostRequest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "accountId": {
            "type": "string"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          }
        }
      },
...
      "CharacterResourcesDto": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "characterId": {
            "type": "string",
            "format": "guid"
          },
          "createdAtUtc": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAtUtc": {
            "type": "string",
            "format": "date-time"
          },
          "healthCurrent": {
            "type": "integer",
            "format": "uint64"
          },
          "healthMaximum": {
            "type": "integer",
            "format": "uint64"
          }
        }
      },
      "CharacterResourcesSearchResponse": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "continuationToken": {
            "type": "string",
            "format": "guid",
            "nullable": true
          },
          "records": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CharacterResourcesDto"
            }
          }
        }
      },
      "CharacterResourcesPatchRequest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "healthCurrent": {
            "type": "integer",
            "format": "uint64",
            "nullable": true
          },
          "healthMaximum": {
            "type": "integer",
            "format": "uint64",
            "nullable": true
          }
        }
      },
      "CharacterResourcesPutRequest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "healthCurrent": {
            "type": "integer",
            "format": "uint64"
          },
          "healthMaximum": {
            "type": "integer",
            "format": "uint64"
          }
        }
      }
    }
  }
}